#6811 - Refactoring
