+++ sormas-backend/src/main/java/de/symeda/sormas/backend/travelentry/services/BaseTravelEntryService.java
@@ -1,0 +1,42 @@
+package de.symeda.sormas.backend.travelentry.services;
+
+import javax.ejb.EJB;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.From;
+import javax.persistence.criteria.Predicate;
+
+import de.symeda.sormas.backend.common.AbstractCoreAdoService;
+import de.symeda.sormas.backend.travelentry.TravelEntry;
+import de.symeda.sormas.backend.travelentry.TravelEntryJurisdictionPredicateValidator;
+import de.symeda.sormas.backend.travelentry.TravelEntryQueryContext;
+import de.symeda.sormas.backend.user.User;
+import de.symeda.sormas.backend.user.UserService;
+
+public class BaseTravelEntryService extends AbstractCoreAdoService<TravelEntry> {
+
+	@EJB
+	protected UserService userService;
+
+	public BaseTravelEntryService() {
+		super(TravelEntry.class);
+	}
+
+	@Override
+	public Predicate createUserFilter(CriteriaBuilder cb, CriteriaQuery cq, From<?, TravelEntry> travelEntryPath) {
+		return inJurisdictionOrOwned(new TravelEntryQueryContext(cb, cq, travelEntryPath));
+	}
+
+	public Predicate inJurisdictionOrOwned(TravelEntryQueryContext qc) {
+		final User currentUser = userService.getCurrentUser();
+		return TravelEntryJurisdictionPredicateValidator.of(qc, currentUser).inJurisdictionOrOwned();
+	}
+
+	public Predicate createDefaultFilter(CriteriaBuilder cb, From<?, TravelEntry> root) {
+		return cb.isFalse(root.get(TravelEntry.DELETED));
+	}
+
+	protected Predicate createUserFilter(TravelEntryQueryContext travelEntryQueryContext) {
+		return inJurisdictionOrOwned(travelEntryQueryContext);
+	}
+}
